% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbsBayesGflasso.R
\name{bfl}
\alias{bfl}
\title{Bayesian fused lasso}
\usage{
bfl(x, y, lambda2, maxiter = 1e+05, burnin = 30000, scaled = T,
  method = "mode_kde")
}
\arguments{
\item{x}{is a matrix of order n x p where n is number of observations and p is number of predictor variables.}

\item{y}{y is a vector of response variable of order n x 1.}

\item{lambda2}{The hyper-parameter control the degree of fusion.}

\item{maxiter}{Gibbs Sampling Number of Iterations.}

\item{burnin}{Number of burnin.}

\item{scaled}{Flag whether to standardize.}

\item{method}{Estimation method from ex posterior distribution sample.}
}
\description{
Bayesian fused lasso
}
