% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebic.R
\name{ebic}
\alias{ebic}
\title{EBIC calculation.}
\usage{
ebic(x, y, beta, sigma2, type = "fusedlasso")
}
\arguments{
\item{x}{is a matrix of order n x p where n is number of observations and p is number of predictor variables.}

\item{y}{y is a vector of response variable of order n x 1.}

\item{beta}{Regression coefficient.}

\item{sigma2}{Dispersion parameter.}

\item{type}{Estimation method from ex posterior distribution sample.}
}
\description{
EBIC calculation.
}
