\name{neggfl-package}
\alias{neggfl-package}
\alias{neggfl}
\docType{package}
\title{
\packageTitle{neggfl}
}
\description{
\packageDescription{neggfl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{neggfl}
\packageIndices{neggfl}
Execute NEG generalized fused lasso.
}
\author{
\packageAuthor{neggfl}

Maintainer: \packageMaintainer{neggfl}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ Bayes sparse modeling, NEG prior distribution }
\examples{
# ~~ Each example is described in the function help ~~
}
