\name{neggfl-package}
\alias{neggfl-package}
\alias{neggfl}
\docType{package}
\title{
\packageTitle{neggfl}
}
\description{
\packageDescription{neggfl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{neggfl}
\packageIndices{neggfl}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{neggfl}

Maintainer: \packageMaintainer{neggfl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
}
